[native] handle missing neynar config in registration
Summary: devs might not have the neynar.json file required to construct the neynar client. if the client doesn't exist, we use the fallback function to make the getFCAvatarURLs call.

Test Plan: removed my neynar.json file and was able to register successfully. as expected, there was no farcaster avatar option available despite connecting my farcaster account. this is because the neynar client cannot be constructed without the json file.

Reviewers: ashoat, kamil

Reviewed By: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D13492