[lib][native][web] Use threadTypeIsPersonal where possible
Summary:
This diff is a no-op. I want to reduce the number of places we use `GENESIS_PERSONAL` in the codebase, so that the only results for `git grep GENESIS_PERSONAL` are places where we need to specifically consider `GENESIS_PERSONAL` rather than general places we are considering personal threads.

Depends on D13550

Test Plan: Careful code inspection

Reviewers: tomek, kamil, inka

Differential Revision: https://phab.comm.dev/D13551