[lib] Update useUserProfileThreadInfo to consider thick threads
Summary: If a thick thread exists with the user whose profile we are viewing, we should display it.

When there are multiple, we'll select the oldest thread (by creation time). This matches behavior of `pendingToRealizedThreadIDsSelector` and `useExistingThreadInfoFinder`.

Depends on D13532

Test Plan: Tested viewing user profile on both web and native with a user with whom the viewer has a thick thread

Reviewers: tomek, kamil, inka