[lib] Simplify useExistingThreadInfoFinder
Summary:
Since D13404, `pendingToRealizedThreadIDsSelector` will return the same thing when queried with a thin thread versus a thick thread.

As such, there's no longer a reason to check for both separately here.

Separately, I also got rid of the ternary at the end in favor of an early exit.

Depends on D13533

Test Plan: Careful reading of the code, as well as being tested with the rest of the stack, where I test all kinds of ways of creating a thread

Reviewers: tomek, kamil, inka

Differential Revision: https://phab.comm.dev/D13534