[terraform] enable multi-AZ for AWS RDS instance
Summary:
Still in the process of testing fallback when upgrading instance size, however putting this diff up for now

This will create a replica of our existing aws rds instance and store it in another availability zone. If our primary aws rds instance fails, queries will be directed to the replica.

This should be applied before we do an instance upgrade as to ensure we minimize downtime.

Test Plan:
terraform apply on self-host set up

Reviewers:
ashoat,
varun,
bartek

Subscribers: