[keyserver] Stop creating GENESIS_PRIVATE thread on account creation
Summary:
This addresses [ENG-9411](https://linear.app/comm/issue/ENG-9411/disable-genesis-private-thread-creation-on-account-creation), although the following diffs are necessary for a related follow-up: making it possible to create a `PRIVATE` thread.

Depends on D13534

Test Plan: Confirm that when a new account is registered, a `GENESIS_PRIVATE` thread is not created for them.

Reviewers: tomek, kamil, inka

Reviewed By: inka

Differential Revision: https://phab.comm.dev/D13535