[lib] Including pending thread for viewer in useThreadListSearch
Summary:
Before this work, we always had a `GENESIS_PRIVATE` thread for the viewer, so we had no reason to support pending private threads in `useThreadListSearch`.

Now we would like to support that. We'll show a pending private thread only if no private thread exists.

Depends on D13536

Test Plan: In combination with the rest of the stack, test creating a `PRIVATE` thread via thread list search on both web and native

Reviewers: tomek, kamil, inka

Differential Revision: https://phab.comm.dev/D13537