[lib] Support creating private thread from user profile
Summary:
Before this diff, when no private thread existed, we wouldn't show a chat link in the user's own user profile. We didn't need to consider such a scenario since we always created a `GENESIS_PRIVATE` thread on account creation.

This diff changes behavior such that if `!oldestPrivateThreadInfo` (no private chat exists), we no longer return `null` from `useUserProfileThreadInfo`, and instead continue on to the `createPendingPersonalOrPrivateThread` call.

Depends on D13538

Test Plan: In combination with the rest of the stack, test creating a `PRIVATE` thread via thread list search on both web and native

Reviewers: tomek, kamil, inka

Differential Revision: https://phab.comm.dev/D13539