[lib] Don't prefix keyserver-owned holders with device ID
Summary:
When uploading media metadata to keyserver, we shouldn't prefix holders with device ID because the holder's owner is actually the keyserver, not the uploading device.
Modified logic to only add prefix if not uploading metadata to keyserver.

Test Plan: Uploaded media to a thin thread and thick thread. The latter had deviceID prefix, the former did not.

Reviewers: kamil, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13517