[lib] Update pending thread type based on selected members
Summary:
Every time a list of selected users change check whether they support thick threads and update the type accordingly. Checking the support is an asynchronous operation which causes a lot of changes in other places.

https://linear.app/comm/issue/ENG-9393/update-pending-thread-type-based-on-selected-members

Depends on D13489

Test Plan:
Modify `useUserSupportThickThread` so that it returns false for some users. Check if the thread type is determined correctly on native and web.

Tested following scenarios

  # Searching for a user on native with useUserSupportThickThread modified to return false for some users
  # Searching for a user on native without modifications to useUserSupportThickThread
  # Searching for a user on web with useUserSupportThickThread modified to return false for some users
  # Searching for a user on web without modifications to useUserSupportThickThread
  # Composing a thread on native with useUserSupportThickThread modified to return false for some users
  # Composing a thread on native without modifications to useUserSupportThickThread
  # Composing a thread on web with useUserSupportThickThread modified to return false for some users
  # Composing a thread on web without modifications to useUserSupportThickThread

For each of these verified that the thread type is displayed correctly and that sending a message results in the creation of the correct thread type.

Reviewers: kamil, inka, ashoat

Reviewed By: kamil, ashoat

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13519