[lib] Update pending thread type based on selected members
Summary:
Every time a list of selected users change check whether they support thick threads and update the type accordingly. Checking the support is an asynchronous operation which causes a lot of changes in other places.

https://linear.app/comm/issue/ENG-9393/update-pending-thread-type-based-on-selected-members

Depends on D13489

Test Plan: Modify `useUserSupportThickThread` so that it returns false for some users. Check if the thread type is determined correctly.

Reviewers: kamil, inka

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13519