[native] Fix crash in NUX
Summary:
issue: [[https://linear.app/comm/issue/ENG-9375/button-should-be-registered-with-nuxtipcontext-error | ENG-9375]]
If a dev kills the app while it is showing the NUX, then after they reopen they will get this error: "button should be registered with nuxTipContext"
This happens because for dev we remember the navigation state. When the navigation state is rehydrated, the buttons have not yet been mounted, so they are not registered with the context, resulting in the error.
Because this is for dev only, and because we expect the NUX tips to only be shown at the start of the app over the chat list, I think we should just reset the nav state if NUX tips are present.

Test Plan:
Created a new user -> saw the NUX tips -> navigated to a tip that points to a button -> closed the app -> opened the app
Before this resulted in an error on the screen. After this change the error doesn't appear. The user sees the chat list.

Reviewers: tomek, kamil, ashoat

Reviewed By: ashoat

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13520