[native] Increase WorkerThread MPMCQueue capacity to 500
Summary:
In [ENG-9416](https://linear.app/comm/issue/ENG-9416/account-deletion-broken-on-build-408), we encountered a scenario where the crypto `WorkerThread` gets filled up and fails to enqueue new tasks.

We have recently been increasing our use of the crypto thread, and in some cases we can enqueue a lot of tasks. I think it makes sense to increase this number.

Test Plan: I haven't directly tested this. I think the change is safe because I'm just increasing a number. I'll make sure that the app builds and runs on Android before landing

Reviewers: varun, will

Subscribers: