[native] Catch when initializeCryptoAccount fails to schedule on crypto thread
Summary:
In [ENG-9416: Account deletion broken on build 408](https://linear.app/comm/issue/ENG-9416/account-deletion-broken-on-build-408), the app crashes because a task on a `WorkerThread` throws an uncaught exception while trying to schedule a task on the crypto `WorkerThread`.

This doesn't happen when the JS thread tries to schedule a task on a `WorkerThread`, as any C++ exceptions on the JS thread are caught and converted to JS exceptions.

The only place we schedule onto the crypto `WorkerThread` from another `WorkerThread` is in `initializeCryptoAccount`. Adding a `try-catch` here allows us to avoid this crash.

Test Plan: I tested using [this patch](https://gist.github.com/Ashoat/93225b247a41cba11cf3c87c9c8f27df) and confirmed that the exception doesn't crash the app

Reviewers: will, varun

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D13525