[native] Combine invokeAsync calls at the end of initializeCryptoAccount
Summary:
I think these were accidentally split up in D11232. In the first revision of that diff, it made sense to separate them. By the final revision, it no longer made sense, but the split was left there.

Depends on D13526

Test Plan: Confirm that the app builds and that `initializeCryptoAccount` still executes successfully

Reviewers: varun, will

Reviewed By: will

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D13527