[lib] Hide initialThreadState.memberIDs from GENESIS
Summary:
We hide the members list of GENESIS. For similar reasons, we should hide the list of initial members.

This addresses [ENG-9430](https://linear.app/comm/issue/ENG-9430/hide-initialthreadstatememberids-list-on-genesis); more context there.

Test Plan: I added an `|| true`, and then logged out and back in on `web` and confirmed that all thread creation robotext were missing "and added A and B" at the end

Reviewers: varun, will

Reviewed By: will

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D13528