[identity] Validate username prefix queries with identity search
Summary: This validates whether username prefixes are valid using regex. If not, we return an empty vector of results

Test Plan: Ended up testing this on staging with a local iOS simulator app connecting to staging identity and running queries. No changes in functionality. Tested with special characters, regular usernames, and ethereum addresses.

Reviewers: varun, bartek

Reviewed By: bartek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D13529