[lib] Don't display notFriendNotice for the viewer
Summary:
Prior to the changes in D13540, `usePotentialMemberItems` never got the viewer as an input, so it didn't need to consider it.

It currently sets `notFriendNotice` for the viewer, which doesn't make sense. While it's true that the viewer is not a friend with themselves, we have no reason to display that notice to the user.

Depends on D13541

Test Plan: Confirm that `notFriendNotice` no longer appears for the viewer (the viewer's entry in the list of users that can be added to the chat)

Reviewers: tomek, kamil, inka

Reviewed By: tomek

Differential Revision: https://phab.comm.dev/D13542