[native][web] Only show viewer in chat composer results when empty
Summary:
If the user has selected somebody other than themselves to chat with, we'll exclude the viewer from the results.

In other words, we'll only show the viewer as an option to themselves when the list is empty.

We're doing this because we want to treat this case similarly to when the viewer selects a non-friend: instead of letting them include themselves in the list, we will just open directly to the `PRIVATE` if the user select themselves at the start (D13545), and not list them as an option if they select somebody else at the start (this diff).

Depends on D13542

Test Plan: Confirm that if the viewer adds somebody to the list of pending members, they can no longer see themselves in the list of potential pending members

Reviewers: tomek, kamil, inka

Reviewed By: tomek

Differential Revision: https://phab.comm.dev/D13543