[native][web] Go to PRIVATE chat when the user selects themselves in chat composer
Summary: We'll handle this just like when the user selects a non-friend: we'll just open the chat directly.

We we want to treat this case similarly to when the viewer selects a non-friend: instead of letting them include themselves in the list, we will just open directly to the `PRIVATE` if the user select themselves at the start (this diff), and not list them as an option if they select somebody else at the start (D13543).

Dpeneds on D13544

Test Plan: On both `native` and `web`, confirm that when the viewer selects themselves, it opens up the pending thread directly

Reviewers: tomek, kamil, inka

Subscribers: