[lib] Make sure we set isViewer in useExistingThreadInfoFinder
Summary: When we call `createPendingThread` in `useExistingThreadInfoFinder`, we need to include `isViewer` in order to make sure a `PRIVATE` thread is rendered as eg. "ashoat" instead of "just you". This is because the naming is based on `ifJustViewer`, which only applies if the members list only includes the viewer.

Depends on D13546

Test Plan: Confirm that I see eg. "ashoat" instead of "just you" for `PRIVATE` thread `uiName`

Reviewers: tomek, kamil, inka

Subscribers: