[web] Hide members option in ThreadMenu for all private and personal threads
Summary:
We currently only hide this menu option for `threadTypes.GENESIS_PERSONAL`. I think we should match behavior for `threadTypes.PERSONAL`, and I also figure we can hide the menu option for private threads too (`threadTypes.PRIVATE` and `threadTypes.GENESIS_PRIVATE`, which are for a user chatting with themselves).

Depends on D13549

Test Plan: Confirm that members options doesn't appear in `ThreadMenu` on `web` when viewing a `PRIVATE` thread

Reviewers: tomek, kamil, inka

Differential Revision: https://phab.comm.dev/D13550