[web] Use PRIVATE instead of GENESIS_PRIVATE in baseThreadInfo passed to useExistingThreadInfoFinder
Summary:
When `searching: true`, `useExistingThreadInfoFinder` doesn't actually care about `baseThreadInfo.type`, as long as it passes `threadTypeCanBePending`. The only exception would be if the viewer is logged out, in which case the calling component should be in the process of being unmounted anyways.

Given that it doesn't matter, I'd prefer to use `PRIVATE` here instead of `GENESIS_PRIVATE`.

`native` changes aren't necessary because they were handled in D13253.

Depends on D13552

Test Plan: Careful code inspection, as well as the testing done in the rest of the stack: on both native and web, tested `PRIVATE` thread creation end-to-end via thread list search, chat composer, and user profile. Confirmed that a peer (same user device) receives the DM messages

Reviewers: tomek, kamil, inka

Subscribers: