[web] send OTKs on v1 login
Summary:
Depends on D13574

Same as previous diff but for web. We want to send OTKs on login since the device ID will be new

https://linear.app/comm/issue/ENG-9440/clients-should-include-otks-on-first-login

Test Plan:
logged in and out multiple times on web. confirmed that the device ID changed in device list and that the OTKs were added (and removed on logout) for each device

tested both password and wallet login

Reviewers: bartek, will, ashoat

Reviewed By: bartek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D13575