[lib] Fix SQLite migrations
Summary:
The migrations are broken because we weren't using the right shape of the specification. This diff fixes it and introduces necessary type checks.

https://linear.app/comm/issue/ENG-9467/redux-migrations-are-broken

Test Plan: Tested this by creating a new migration on the web where a new draft is created. Run the app, and reopened it. Checked if the SET_CLIENT_DB_STORE action contains the draft.

Reviewers: kamil, ashoat, atul

Differential Revision: https://phab.comm.dev/D13571