[lib] Bump detect-browser version
Summary: We're using an old version that has more limited browsers support. According to the release notes https://github.com/DamonOehlman/detect-browser/releases/tag/v5.0.0 the major version was increased because of a breaking change that removed support for one browser.

https://linear.app/comm/issue/ENG-9477/typeerror-cannot-read-properties-of-null-reading-os-on-landing

Test Plan:
Tested on iOS that the correct version is detected in the responder.

Reviewers: kamil, inka