[keyserver] Handle null browser detection result
Summary:
When a browser can't be detected we need to make sure that we don't crash. Handling this case in the same way as if it was a web browser probably makes the most sense.

https://linear.app/comm/issue/ENG-9477/typeerror-cannot-read-properties-of-null-reading-os-on-landing

Depends on D13587

Test Plan: Replaced `detectionResult` with `null` and checked that we're handling it as if it was a web browser.

Reviewers: kamil, inka