[lib] Update the dev invite link URL
Summary: Update the dev invite link URL so that it works for the default configuration. Making it correct in all the cases is hard because we would need the `webapp_url.json` config that isn't available in `lib`.

Test Plan: Open the dev app and check if the link is displayed correctly and works.

Reviewers: kamil, inka

Reviewed By: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13589