[keyserver] Show the most recent thread when opening the web app without providing a thread ID
Summary:
The most recent thread can be thick and the keyserver doesn't know about it. In this case we can rely on the app to compute the `activeChatThreadID` instead of the keyserver. In `validateStateAndQueueOpsProcessing` we're handling an empty `state.navInfo.activeChatThreadID` by calling `mostRecentlyReadThreadSelector`. This means that we can simply stop handling empty `activeChatThreadID` on the keyserver side.

https://linear.app/comm/issue/ENG-9126/opening-a-link-to-the-chat-doesnt-open-the-most-recent-thick-thread

Test Plan:
Made sure that the most recent thread is thick. Opened `/webapp` and `/webapp/chat` and verified that the most recent thread is shown.
Updated the state so that a thin thread is the most recent and performed the same test.
Made sure that opening a link to a thin and to a thick thread still opens it even if it isn't the most recent.

Reviewers: kamil, inka

Reviewed By: inka

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13590