[lib] Remove !threadInfos check from BaseAutoJoinCommunityHandler
Summary:
Ginsu added this check in [this revision](https://phab.comm.dev/D12065?vs=40275&id=41343#toc) of the diff that first introduced this code.

My understanding is that `threadInfos` should always be set, and that's how we treat it elsewhere in the codebase. I think it's safe to remove this check.

Test Plan: Flow, and also checking with Ginsu about why he added the check

Reviewers: varun

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D13626