[native] Avoid updating relationships before the previous call finished
Summary:
Make sure we don't update the relationship before the previous call ended. Currently, it is possible to click a button multiple times in a row which results in multiple thick threads being created.

https://linear.app/comm/issue/ENG-9508/adding-new-friends-can-result-in-creating-multiple-thick-threads

Test Plan: Click the `save` button multiple times and check if only one thread was created.

Reviewers: kamil, inka

Reviewed By: inka

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13632