[web] Convert ComposedMessage to function component
Summary:
The class component just had a single render function, so it was an easy migration.

This diff should be a clean refactor. I'll add some memoization in the next diff.

Depends on D13671

Test Plan: Flow, and playing around with web app in local environment with robotext, multimedia, and text messages, and sidebars and reactions on each

Reviewers: tomek

Subscribers: