[lib] move calling `getThreadUpdatesForNewMessages` to `reduceThreadInfos`
Summary:
This is a fix for [ENG-9453](https://linear.app/comm/issue/ENG-9453/reply-count-in-sidebars-can-be-incorrect-on-a-sending-device).

Alternative, is introducing another queue in `useSendComposableDMOperation` to make sure we're always calculating replies count on newest thread info, but I prefer this solution and it's cleaner.

Test Plan:
Send a lot of messages and checking replies count - did test multiple times as there is no single way to reproduce it.

Reviewers: tomek