[native] Remove TooltipInlineEngagement from RobotextMessageTooltipButton
Summary:
This was introduced first in D4825 to solve a problem having to do with `InlineEngagement` appearing over the message.

But for robotext messages, we don't overlap the display of `InlineEngagement` with the message. As such, this problem doesn't exist for robotext messages, and doesn't need to be solved.

Depends on D13675

Test Plan: Remove `TooltipInlineEngagement` from text messages and multimedia messages. Observe that the issue described in D4825 appears now. Do the same for robotext messages, and confirm that the issue does not appear.

Reviewers: tomek

Reviewed By: tomek

Differential Revision: https://phab.comm.dev/D13676