[lib][native][web] Introduce chatMessageItemHasEngagement
Summary:
This diff addresses a bug where `MultimediaMessageTooltipButton` and `TextMessageTooltipButton` on `native` were only rendering `TooltipInlineEngagement` when a sidebar existed, but not when a reaction existed.

`TooltipInlineEngagement` is meant to be rendered exactly when `InlineEngagement` is, so these checks should be identical. It was introduced in D4825, which was (slightly) before reactions were implemented (you can see reactions in the test plan, but diff comments review that it was hacked as a demo, and reactions at the time had not been landed yet).

Depends on D13677

Test Plan: Confirm visually (on iOS simulator) that `TooltipInlineEngagement` gets rendered over cases where a message has a reaction but not a sidebar

Reviewers: tomek

Subscribers: