[keyserver] factor out identity auth and saveIdentityInfo from verifyUserLoggedIn
Summary:
we'll use the authAndSaveIdentityInfo function in keyserver.js where we currently call verifyUserLoggedIn. this is because we care about how the identity info was fetched. if it was fetched directly from the identity service, there's no need to call syncPlatformDetails. however, if it fetched from mariadb, we should attempt to sync.

Depends on D13665

Test Plan: tested in next diff

Reviewers: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D13666