[native] Don't create space for author name in RobotextMessageTooltipModal
Summary:
The logic for navigate to `RobotextMessageTooltipModal` is copied from similar logic for `TextMessageTooltipModal` and `MultimediaMessageTooltipModal`.

When the viewer selects a text or multimedia message that somebody else sent, we display the author's name. However, we don't do the same for robotext.

The code being removed here would offsite the modal to give space for the author's name. Since we don't display it for robotext, I'm removing the offset here.

Depends on D13679

Test Plan: Confirm that the tooltip doesn't appear strangely higher for non-viewer-created robotext

Reviewers: tomek

Subscribers: