[lib] Introduce MessageSpec.mergeIntoPrecedingRobotextMessageItem
Summary:
This function will allow us to merge `RobotextMessageInfo`s into preceding (older) `RobotextChatMessageInfoItem`s.

The merging is implemented on the spec of the newer `RobotextChatMessageInfoItem`.

Depends on D13686

Test Plan: Tested in combination with the following diffs, where I construct `RobotextChatMessageInfoItem` with an array of `messageInfos` to represent combined messages ([ENG-9559](https://linear.app/comm/issue/ENG-9559/combine-some-kinds-of-robotext-when-they-appear-together)

Reviewers: tomek

Reviewed By: tomek

Differential Revision: https://phab.comm.dev/D13687